--- src/main/scala/de/johoop/jacoco4sbt/JaCoCoRuntime.scala
@@ -1,23 +1,0 @@
-/*
- * This file is part of jacoco4sbt.
- *
- * Copyright (c) Joachim Hofer & contributors
- * All rights reserved.
- *
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-
-package de.johoop.jacoco4sbt
-
-import sbt._
-import Keys._
-import org.jacoco.core.runtime.RuntimeData
-import org.jacoco.core.runtime.LoggerRuntime
-
-private[jacoco4sbt] trait JaCoCoRuntime {
-  val runtimeData = new RuntimeData
-  val runtime = new LoggerRuntime
-}
