+++ src/main/scala/org/scalasbt/jacoco/report/JacocoSourceSettings.scala
@@ -1,0 +1,15 @@
+/*
+ * This file is part of jacoco4sbt.
+ *
+ * Copyright (c) Joachim Hofer & contributors
+ * All rights reserved.
+ *
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.scalasbt.jacoco.report
+
+case class JacocoSourceSettings(tabWidth: Int = 2, fileEncoding: String = "utf-8")
